# api/public/openapi.yaml
openapi: 3.0.3
info:
  title: Horus API
  version: "0.1.0"
servers:
  - url: http://localhost:8000
paths:
  /api/login:
    post:
      summary: Login
      description: >
        Login local com token interno (Sanctum).

        O endpoint é extensível para múltiplos canais no futuro, mas atualmente o canal suportado é `internal`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - channel
                - payload
              properties:
                channel:
                  type: string
                  enum: [internal]
                  example: internal
                payload:
                  type: object
                  description: Payload específico do canal.
                  properties:
                    email:
                      type: string
                      format: email
                      example: internal@example.com
                    password:
                      type: string
                      example: "********"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                        example: 1|8fG09Kcxc1gnGqGG064tWT8XtSJLDJi1GHYpjKFG
                      name:
                        type: string
                        example: John Doen
                      email:
                        type: string
                        example: johndoe@example.com
        "422":
          description: Erro de validação
  /api/me:
    post:
      summary: Valida sessão/token
      description: >
        Valida o Bearer token (Sanctum) e retorna os dados do usuário autenticado.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Usuário logado.
        "401":
          description: Não autenticado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Usuário não autenticado.
  /api/logout:
    post:
      summary: Encerra sessão (revoga token atual)
      description: >
        Revoga o token de acesso Bearer usado na requisição (Sanctum).
        Após o logout, o mesmo token não deve ser aceito em rotas protegidas.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Sessão encerrada.
                  data:
                    type: object
                    properties:
                      ok:
                        type: boolean
                        example: true
        "401":
          description: Não autenticado
  /api/users:
    get:
      summary: Lista usuários
      description: Paginação por `page` e `per_page` (máx. 100). Requer autenticação Sanctum.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - in: query
          name: per_page
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: OK
        "401":
          description: Não autenticado
        "422":
          description: Erro de validação
    post:
      summary: Cria usuário
      description: Cria um registro na tabela `users` (gestão interna).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password, password_confirmation]
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                password_confirmation:
                  type: string
      responses:
        "201":
          description: Created
        "401":
          description: Não autenticado
        "422":
          description: Erro de validação
  /api/users/{user}:
    parameters:
      - in: path
        name: user
        required: true
        schema:
          type: integer
    get:
      summary: Detalhe do usuário
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: format
          schema:
            type: string
            enum: [json]
      responses:
        "200":
          description: OK
        "401":
          description: Não autenticado
        "404":
          description: Não encontrado
        "422":
          description: Erro de validação
    put:
      summary: Atualiza usuário (substituição)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email]
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                password_confirmation:
                  type: string
      responses:
        "200":
          description: OK
        "401":
          description: Não autenticado
        "404":
          description: Não encontrado
        "422":
          description: Erro de validação
    patch:
      summary: Atualiza usuário (parcial)
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                password_confirmation:
                  type: string
      responses:
        "200":
          description: OK
        "401":
          description: Não autenticado
        "404":
          description: Não encontrado
        "422":
          description: Erro de validação
    delete:
      summary: Remove usuário
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: dry_run
          schema:
            type: boolean
      responses:
        "200":
          description: OK
        "401":
          description: Não autenticado
        "404":
          description: Não encontrado
        "422":
          description: Erro de validação
  /api/logs:
    get:
      summary: Lista logs com filtros e paginação
      description: >
        Realiza busca no OpenSearch e retorna logs paginados.
        Suporta filtros por datas, mensagem, canal, nível, entidade e exceção.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
          example: 1
        - in: query
          name: per_page
          schema:
            type: integer
            minimum: 1
            maximum: 100
          example: 50
        - in: query
          name: sort
          schema:
            type: string
            enum: [received_at, processed_at, created_at]
          example: received_at
        - in: query
          name: order
          schema:
            type: string
            enum: [asc, desc]
          example: desc
        - in: query
          name: filters[level]
          schema:
            type: string
            enum: [debug, info, notice, warning, error, critical, alert, emergency]
          example: error
        - in: query
          name: filters[channel]
          schema:
            type: string
          example: http
        - in: query
          name: filters[message]
          schema:
            type: string
          example: exception
        - in: query
          name: filters[entity_name]
          schema:
            type: string
          example: order
        - in: query
          name: filters[entity_id]
          schema:
            type: string
          example: "123"
        - in: query
          name: filters[has_exception]
          schema:
            type: boolean
          example: true
        - in: query
          name: filters[received_at][from]
          schema:
            type: string
            format: date
          example: 2026-04-01
        - in: query
          name: filters[received_at][to]
          schema:
            type: string
            format: date
          example: 2026-04-30
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        tracking_id:
                          type: string
                          example: 01JT18J1LQF5D6Q8XQ0Y2AZ123
                        level:
                          type: string
                          example: error
                        message:
                          type: string
                          example: Unexpected exception while processing order
                        channel:
                          type: string
                          nullable: true
                          example: http
                        entity_name:
                          type: string
                          nullable: true
                          example: order
                        entity_id:
                          type: string
                          nullable: true
                          example: "123"
                        exception:
                          type: object
                          nullable: true
                        received_at:
                          type: string
                          example: 2026-04-29T09:00:00+00:00
                  meta:
                    type: object
                    properties:
                      total:
                        type: integer
                        example: 1
                      page:
                        type: integer
                        example: 1
                      per_page:
                        type: integer
                        example: 50
    post:
      summary: Recebe um log para processamento assíncrono
      description: >
        Valida o Bearer token no servidor externo usando `/auth/me`,
        valida o payload, gera `tracking_id` e publica o log na fila de processamento.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - level
                - message
                - source
                - environment
              properties:
                level:
                  type: string
                  enum: [debug, info, notice, warning, error, critical, alert, emergency]
                  example: error
                message:
                  type: string
                  example: Unexpected exception while processing order
                context:
                  type: object
                  additionalProperties: true
                entity_name:
                  type: string
                  nullable: true
                  example: order
                entity_id:
                  type: string
                  nullable: true
                  example: "123"
                source:
                  type: string
                  example: billing-api
                environment:
                  type: string
                  example: production
                channel:
                  type: string
                  nullable: true
                  example: http
                request_id:
                  type: string
                  nullable: true
                  example: req-123
                trace_id:
                  type: string
                  nullable: true
                  example: trace-123
                user_id:
                  type: string
                  nullable: true
                  example: user-123
                ip_address:
                  type: string
                  nullable: true
                  example: 10.0.0.10
                user_agent:
                  type: string
                  nullable: true
                  example: Mozilla/5.0
                exception:
                  type: object
                  nullable: true
                  properties:
                    class:
                      type: string
                      nullable: true
                      example: RuntimeException
                    message:
                      type: string
                      nullable: true
                      example: Failure
                    file:
                      type: string
                      nullable: true
                      example: /var/www/app/OrderService.php
                    line:
                      type: integer
                      nullable: true
                      example: 88
                    stack_trace:
                      type: string
                      nullable: true
                      example: stack trace content
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Log recebido para processamento
                  tracking_id:
                    type: string
                    example: 01HXABC123
        "401":
          description: Não autenticado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Token inválido ou não autorizado.
        "422":
          description: Erro de validação

  /api/logs/summary:
    get:
      summary: Agregações para dashboard
      description: >
        Busca única (`size` 0) com agregações term e histogram sobre o índice `logs-*`.
        Sobrescreve o intervalo em `filters.received_at` com a janela de últimos `histogram_days` (UTC)
        combinada aos demais filtros enviados.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: histogram_days
          schema:
            type: integer
            minimum: 1
            maximum: 90
          example: 14
          description: Largura em dias do histograma diário e do filtro de recebimento.
        - in: query
          name: filters[level]
          schema:
            type: string
            enum: [debug, info, notice, warning, error, critical, alert, emergency]
        - in: query
          name: filters[channel]
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      total:
                        type: integer
                      with_exception_count:
                        type: integer
                      period:
                        type: object
                        properties:
                          from:
                            type: string
                            format: date
                          to:
                            type: string
                            format: date
                          histogram_days:
                            type: integer
                      by_level:
                        type: array
                        items:
                          type: object
                          properties:
                            key:
                              type: string
                            count:
                              type: integer
                      by_channel:
                        type: array
                        items:
                          type: object
                          properties:
                            key:
                              type: string
                            count:
                              type: integer
                      by_source:
                        type: array
                        items:
                          type: object
                          properties:
                            key:
                              type: string
                            count:
                              type: integer
                      by_environment:
                        type: array
                        items:
                          type: object
                          properties:
                            key:
                              type: string
                            count:
                              type: integer
                      histogram:
                        type: array
                        items:
                          type: object
                          properties:
                            date:
                              type: string
                              example: "2026-04-01"
                            count:
                              type: integer
        "401":
          description: Não autenticado

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT